refactor - #111 auth 패키지에서 user 기능에 관련한 내용 user 패키지로 이관
